Always add WWW-Authenticate header when auth fails
This is required so the user agent can determine what auth
method is supported.

Quote from RFC 2617:
If the origin server does not wish to accept the credentials sent
with a request, it SHOULD return a 401 (Unauthorized) response. The
response MUST include a WWW-Authenticate header field containing at
least one (possibly new) challenge applicable to the requested
resource.